+++ src/test/java/org/stellar/sdk/FeeBumpTransactionTest.java
@@ -4,9 +4,30 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.fail;
 
+import java.math.BigInteger;
+import java.util.ArrayList;
 import org.junit.Test;
+import org.stellar.sdk.operations.InvokeHostFunctionOperation;
 import org.stellar.sdk.operations.PaymentOperation;
+import org.stellar.sdk.xdr.ContractExecutable;
+import org.stellar.sdk.xdr.ContractExecutableType;
+import org.stellar.sdk.xdr.ContractIDPreimage;
+import org.stellar.sdk.xdr.ContractIDPreimageType;
+import org.stellar.sdk.xdr.CreateContractArgs;
 import org.stellar.sdk.xdr.EnvelopeType;
+import org.stellar.sdk.xdr.ExtensionPoint;
+import org.stellar.sdk.xdr.HostFunction;
+import org.stellar.sdk.xdr.HostFunctionType;
+import org.stellar.sdk.xdr.Int64;
+import org.stellar.sdk.xdr.LedgerEntryType;
+import org.stellar.sdk.xdr.LedgerFootprint;
+import org.stellar.sdk.xdr.LedgerKey;
+import org.stellar.sdk.xdr.SignerKey;
+import org.stellar.sdk.xdr.SorobanResources;
+import org.stellar.sdk.xdr.SorobanTransactionData;
+import org.stellar.sdk.xdr.Uint256;
+import org.stellar.sdk.xdr.Uint32;
+import org.stellar.sdk.xdr.XdrUnsignedInteger;
 
 public class FeeBumpTransactionTest {
 
@@ -22,6 +43,12 @@ public class FeeBumpTransactionTest {
                     .destination("GA7QYNF7SOWQ3GLR2BGMZEHXAVIRZA4KVWLTJJFC7MGXUA74P7UJVSGZ")
                     .asset(new AssetTypeNative())
                     .amount("200")
+                    .build())
+            .addOperation(
+                PaymentOperation.builder()
+                    .destination("GA7QYNF7SOWQ3GLR2BGMZEHXAVIRZA4KVWLTJJFC7MGXUA74P7UJVSGZ")
+                    .asset(new AssetTypeNative())
+                    .amount("300")
                     .build())
             .setBaseFee(baseFee)
             .addPreconditions(
@@ -214,5 +241,114 @@ public class FeeBumpTransactionTest {
             createInnerTransaction(Network.PUBLIC));
 
     assertNotEquals(feeBump0, feeBump3);
+  }
+
+  @Test
+  public void testCreateWithBaseFee() {
+    Transaction inner = createInnerTransaction(300);
+    FeeBumpTransaction feeBumpTx =
+        FeeBumpTransaction.createWithBaseFee(
+            "GDQNY3PBOJOKYZSRMK2S7LHHGWZIUISD4QORETLMXEWXBI7KFZZMKTL3", 400, inner);
+
+    assertEquals(
+        feeBumpTx.getFeeSource(), "GDQNY3PBOJOKYZSRMK2S7LHHGWZIUISD4QORETLMXEWXBI7KFZZMKTL3");
+    assertEquals(feeBumpTx.getFee(), 400 * 3);
+    assertEquals(feeBumpTx.getInnerTransaction(), inner);
+  }
+
+  @Test
+  public void testCreateWithFee() {
+    Transaction inner = createInnerTransaction(300);
+    FeeBumpTransaction feeBumpTx =
+        FeeBumpTransaction.createWithFee(
+            "GDQNY3PBOJOKYZSRMK2S7LHHGWZIUISD4QORETLMXEWXBI7KFZZMKTL3", 1100, inner);
+
+    assertEquals(
+        feeBumpTx.getFeeSource(), "GDQNY3PBOJOKYZSRMK2S7LHHGWZIUISD4QORETLMXEWXBI7KFZZMKTL3");
+    assertEquals(feeBumpTx.getFee(), 1100);
+    assertEquals(feeBumpTx.getInnerTransaction(), inner);
+  }
+
+  @Test
+  public void testCreateWithBaseFeeWithSorobanOp() {
+    long sorobanResourceFee = 346546L;
+    KeyPair source =
+        KeyPair.fromSecretSeed("SCH27VUZZ6UAKB67BDNF6FA42YMBMQCBKXWGMFD5TZ6S5ZZCZFLRXKHS");
+
+    Account account = new Account(source.getAccountId(), 2908908335136768L);
+    LedgerKey ledgerKey =
+        LedgerKey.builder()
+            .discriminant(LedgerEntryType.ACCOUNT)
+            .account(
+                LedgerKey.LedgerKeyAccount.builder()
+                    .accountID(
+                        KeyPair.fromAccountId(
+                                "GB7TAYRUZGE6TVT7NHP5SMIZRNQA6PLM423EYISAOAP3MKYIQMVYP2JO")
+                            .getXdrAccountId())
+                    .build())
+            .build();
+    SorobanTransactionData sorobanData =
+        SorobanTransactionData.builder()
+            .resources(
+                SorobanResources.builder()
+                    .footprint(
+                        LedgerFootprint.builder()
+                            .readOnly(new LedgerKey[] {ledgerKey})
+                            .readWrite(new LedgerKey[] {})
+                            .build())
+                    .readBytes(new Uint32(new XdrUnsignedInteger(699)))
+                    .writeBytes(new Uint32(new XdrUnsignedInteger(0)))
+                    .instructions(new Uint32(new XdrUnsignedInteger(34567)))
+                    .build())
+            .resourceFee(new Int64(sorobanResourceFee))
+            .ext(ExtensionPoint.builder().discriminant(0).build())
+            .build();
+
+    CreateContractArgs createContractArgs =
+        CreateContractArgs.builder()
+            .contractIDPreimage(
+                ContractIDPreimage.builder()
+                    .discriminant(ContractIDPreimageType.CONTRACT_ID_PREIMAGE_FROM_ADDRESS)
+                    .fromAddress(
+                        ContractIDPreimage.ContractIDPreimageFromAddress.builder()
+                            .address(
+                                new Address(
+                                        "GB7TAYRUZGE6TVT7NHP5SMIZRNQA6PLM423EYISAOAP3MKYIQMVYP2JO")
+                                    .toSCAddress())
+                            .salt(new Uint256(new byte[32]))
+                            .build())
+                    .build())
+            .executable(
+                ContractExecutable.builder()
+                    .discriminant(ContractExecutableType.CONTRACT_EXECUTABLE_STELLAR_ASSET)
+                    .build())
+            .build();
+    HostFunction hostFunction =
+        HostFunction.builder()
+            .discriminant(HostFunctionType.HOST_FUNCTION_TYPE_CREATE_CONTRACT)
+            .createContract(createContractArgs)
+            .build();
+    InvokeHostFunctionOperation invokeHostFunctionOperation =
+        InvokeHostFunctionOperation.builder().hostFunction(hostFunction).build();
+    Transaction transaction =
+        new Transaction(
+            account.getAccountId(),
+            Transaction.MIN_BASE_FEE,
+            account.getIncrementedSequenceNumber(),
+            new org.stellar.sdk.operations.Operation[] {invokeHostFunctionOperation},
+            null,
+            new TransactionPreconditions(
+                null, null, BigInteger.ZERO, 0, new ArrayList<SignerKey>(), null),
+            sorobanData,
+            Network.TESTNET);
+
+    FeeBumpTransaction feeBumpTransaction =
+        FeeBumpTransaction.createWithBaseFee(
+            "GDQNY3PBOJOKYZSRMK2S7LHHGWZIUISD4QORETLMXEWXBI7KFZZMKTL3", 200, transaction);
+    assertEquals(
+        feeBumpTransaction.getFeeSource(),
+        "GDQNY3PBOJOKYZSRMK2S7LHHGWZIUISD4QORETLMXEWXBI7KFZZMKTL3");
+    assertEquals(feeBumpTransaction.getFee(), 200 * (1 + 1) + sorobanResourceFee);
+    assertEquals(feeBumpTransaction.getInnerTransaction(), transaction);
   }
 }